+++ handlers/main.yml
@@ -5,12 +5,12 @@
 # http://stackoverflow.com/a/35130254
 
 - name: omero-web rewrite omero-web configuration
-  become: yes
+  become: true
   become_user: "{{ omero_web_system_user }}"
   command: "{{ omero_web_basedir }}/config/omero-web-config-update.sh"
 
 - name: omero-web restart omero-web
-  become: yes
+  become: true
   service:
     name: omero-web
     state: restarted
