--- molecule.yml
@@ -1,46 +1,0 @@
----
-dependency:
-  name: galaxy
-  requirements_file: tests/requirements.yml
-
-driver:
-  name: docker
-
-vagrant:
-  platforms:
-    - name: centos7
-      box: centos/7
-  providers:
-    - name: virtualbox
-      type: virtualbox
-      options:
-        memory: 512
-        cpus: 1
-  instances:
-    - name: omero-web
-
-docker:
-  containers:
-  - name: omero-web
-    image: centos/systemd
-    image_version: latest
-    privileged: True
-  - name: omero-web-inactive
-    image: centos
-    image_version: 7
-
-ansible:
-  host_vars:
-    omero-web:
-      omero_web_release: latest
-    omero-web-inactive:
-      # This tests a hypothetical instance where systemd is inactive but might
-      # be activated in future
-      # It also tests this role can be used to install omero.web on a system
-      # that will never have systemd, e.g. a simple docker container
-      omero_web_systemd_start: False
-      omero_web_setup_nginx: False
-      omero_web_release: present
-
-verifier:
-  name: testinfra
