+++ molecule/active/tests/test_default.py
@@ -1,12 +1,22 @@
-import testinfra.utils.ansible_runner
+import os
 import pytest
 import re
 
+import testinfra.utils.ansible_runner
+
 testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
-    '.molecule/ansible_inventory').get_hosts('omero-web')
+    os.environ['MOLECULE_INVENTORY_FILE']).get_hosts('all')
+
 
 OMERO = '/opt/omero/web/OMERO.web/bin/omero'
 VERSION_PATTERN = re.compile('(\d+)\.(\d+)\.(\d+)-ice36-')
+
+
+def test_omero_web_config(Command, Sudo):
+    with Sudo('omero-web'):
+        cfg = Command.check_output("%s config get" % OMERO)
+    assert cfg == (
+        'omero.web.server_list=[["localhost", 12345, "molecule-test"]]')
 
 
 def test_omero_version(Command, Sudo):
