+++ molecule/inactive/Dockerfile.j2
@@ -1,0 +1,1 @@
+../resources/Dockerfile.j2
