+++ molecule/resources/playbook.yml
@@ -1,5 +1,7 @@
+---
 # Install 5.3 on both omero-web omero-web-inactive
-- hosts: all
+- name: Converge
+  hosts: all
 
   pre_tasks:
     # To test the upgrade process without breaking the idempotence check
@@ -15,17 +17,17 @@
     - role: ansible-role-omero-web
       omero_web_config_set:
         omero.web.server_list:
-        - [localhost, 12345, molecule-test]
+          - [localhost, 12345, molecule-test]
       omero_web_release: "5.3"
       when: not molecule_test_omero_web_installed_1.stat.exists
 
   tasks:
     - name: get omero-web version
-      become: yes
+      become: true
       become_user: omero-web
       command: /opt/omero/web/OMERO.web/bin/omero version
       register: molecule_test_omero_web_version
-      changed_when: False
+      changed_when: false
 
     - name: check version
       assert:
@@ -36,9 +38,10 @@
 # omero_web_release is defined in molecule.yml host_vars so:
 # - omero-web: latest: upgraded
 # - omero-web-inactive: present: unchanged at 5.3
-- hosts: all
+- name: Converge upgrade
+  hosts: all
   roles:
     - role: ansible-role-omero-web
       omero_web_config_set:
         omero.web.server_list:
-        - [localhost, 12345, molecule-test]
+          - [localhost, 12345, molecule-test]
