+++ tasks/web-systemd.yml
@@ -2,23 +2,23 @@
 # Setup systemd files
 
 - name: omero web | systemd service
-  become: yes
+  become: true
   template:
     dest: /etc/systemd/system/omero-web.service
-    force: yes
+    force: true
     src: systemd-system-omero-web-service.j2
   notify:
-  - reload systemd
-  - omero-web restart omero-web
+    - reload systemd
+    - omero-web restart omero-web
 
 # Flush handlers to ensure systemd is reloaded
 - name: omero web | flush systemd handlers
   meta: flush_handlers
 
 - name: omero web | enable and start web systemd
-  become: yes
+  become: true
   service:
-    enabled: yes
+    enabled: true
     name: omero-web.service
     state: started
   when: omero_web_systemd_start
